Remove blank file and separate values into their own files
