--- src/ILLink.Shared/TrimAnalysis/ReferenceValue.cs
@@ -1,6 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace ILLink.Shared.TrimAnalysis
-{
-}
